Split blockchain.cpp into class per file.
