+++ Makefile.am
@@ -39,6 +39,8 @@ src_libbitcoin_blockchain_la_SOURCES = \
     src/blockchain.cpp \
     src/checkpoints.cpp \
     src/db_interface.cpp \
+    src/fetch_block.cpp \
+    src/fetch_block_locator.cpp \
     src/organizer.cpp \
     src/orphans_pool.cpp \
     src/transaction_pool.cpp \
@@ -284,6 +286,8 @@ include_bitcoin_blockchain_HEADERS = \
     include/bitcoin/blockchain/checkpoints.hpp \
     include/bitcoin/blockchain/db_interface.hpp \
     include/bitcoin/blockchain/define.hpp \
+    include/bitcoin/blockchain/fetch_block.hpp \
+    include/bitcoin/blockchain/fetch_block_locator.hpp \
     include/bitcoin/blockchain/organizer.hpp \
     include/bitcoin/blockchain/orphans_pool.hpp \
     include/bitcoin/blockchain/simple_chain.hpp \
